move fixture to conftest file
